RELEASE 1.0.02
- Refatoracao de entidades
- Criacao de teste Roupa

+++ server/src/main/java/backend/lavanderia/pedido/controller/RoupaController.java
@@ -64,7 +64,7 @@ public class RoupaController
 		ValidaRoupa.roupa(roupa);
 		
 		buscaRoupa.get().setNomeRoupa(roupa.getNomeRoupa());
-		buscaRoupa.get().setPrazo(Long.valueOf(roupa.getPrazo()));
+		buscaRoupa.get().setPrazo(Integer.valueOf(roupa.getPrazo()));
 		buscaRoupa.get().setValor(roupa.getValor());
 		buscaRoupa.get().setImagem(roupa.getImagem());
